+++ tibiapy/parsers/__init__.py
@@ -1,11 +1,7 @@
 from abc import abstractmethod
 from typing import TypeVar, Generic
 
-from tibiapy.parsers.character import *
-from tibiapy.parsers.news import NewsArchiveParser, NewsParser
-
 T = TypeVar('T')
-
 
 class BaseParser(Generic[T]):